# ENH: Optimize calculate