### Summary

For a set of 30M points (x16 DOF = 500M array elements), `calculate` had an overhead of ~35s beyond calling the objective function. These changes optimize that down to ~13s in the general case and ~5s in special cases.

### Problem

When working with a large (8 sublattice) binary system, which results in points arrays of approximately 500M points at a point density of 1000. calling calculate was very slow.

This PR introduces some optimizations to speed up calculate.

For reference, here is the script that was run:

```python
from pycalphad import calculate, Database

dbf = Database('/Users/brandon/Projects/notebooks/pycalphad/AB8SL_v4.TDB')

from pycalphad import Model
mod = Model(dbf, ['A', 'B'], 'FCC')  # save time initializing models on each call
# warm the cache
calc_res = calculate(dbf, ['A', 'B'], 'FCC', pdens=1000, model=mod)
% timeit calc_res = calculate(dbf, ['A', 'B'], 'FCC', pdens=1000, model=mod)
```

### Details

Running this on develop master gives:

`1min 15s ± 783 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`, where ~40s comes from calling the `obj` method of a PhaseRecord. Here I am assuming that this cannot be changed.

The main changes are 

1. Update `dof` array to tranpose before, rather than after, where the concatenated arrays line up better in memory. Leads to 

    `58.7 s ± 2.74 s per loop (mean ± std. dev. of 7 runs, 1 loop each)`

2. Changing to append the NaNs to extend the points array, rather than preallocate a whole array of them and then copy the points over. Both methods have to copy the points, so it's not the most efficient, but this is a little faster because it saves from allocating `np.size(points)` of `np.nan`. The ideal situation (as commented) would be to give `_compute_phase_values` the NaN-padded points array in the first place, then slice that up (if possible) inside the function, preventing any copying of memory. This optimization gives

    `53.1 s ± 4.65 s per loop (mean ± std. dev. of 7 runs, 1 loop each)`

3. For the hot path, such as single phase calculations and other special cases, where the current internal degrees of freedom is the same as the maximum degrees of freedom, we can avoid any extending/copying completely. Gives

    `46.6 s ± 884 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

    which is close to the objective function time. 

Except for the additional overhead of the if statement checking the array shapes, this should across the board be cheaper, even for small points arrays, and should allow us to handle more sublattice and multicomponent systems more gracefully.