@bocklund 2018-06-03
For the record:

After this optimization, assuming a precompiled Model that doesn't have to be initialized in `equilibrium`, the breakdown of an `equilibrium` call is ~98% in dask calculating things.

Of that time, 92.3% is spent in `calculate`. Virtually all of `calculate`is spent in `_compute_phase_values`, where there it's 65.2% calling the objective and 20.6% array allocation/copying (`np.concatenate`, `np.full`), 11% internal logic, and the rest small bits. 

So in order to be more performant for bigger systems we need to (IMO) do the following.

1. Find ways to speed up the objective function (@richardotis suggested reducing branching at Model creation)
2. Reduce array allocation/copying
3. Continue to speed up internal logic in `_compute_phase_values`