@richardotis 2018-06-03
Thanks for this work. If you want to eliminate some more array allocations, I would also look at `calculate.py/_sample_phase_constitution()` and also `utils.py/endmember_matrix()`. I think it is possible to change them to accept a `maximum_internal_dof` argument to ensure all the points arrays are pre-padded. Note that you may need to then tweak some nan-checking logic like `calculate.py:116`.