@bocklund 2018-06-03
My reasoning for not doing looking into padding here was that I think `phase_record.obj(phase_output, dof)` requires the inputs to be C-ordered blocks of arrays in memory, but pre-padding would create more columns and slicing would create a view of that array that is not continuous in memory.

Will passing in the view still work?