@richardotis 2018-06-03
There's a codegen'd Cython wrapper over the raw function which I think will check the dof array shape at call time and do the pointer arithmetic, so all the array indexing operations should safely ignore any `nan` padding at the end of each row.

We should probably also consider Cythonization of `_compute_phase_values` given the huge number of array operations in the hot path.