@richardotis 2018-06-03
Oh right, we bypass the Cython wrapper function. Hmm, I'll have to take a peek at the raw C code generated again to figure out if it's safe.