@richardotis 2018-06-03
I bet we'd see some nice speedups with a Cython pass on `_compute_phase_values`. We could then use [OpenMP](http://docs.cython.org/en/latest/src/userguide/parallelism.html) to chunk the rows of big point arrays and perform the computation in parallel.