@richardotis 2018-06-03
@bocklund I think we could probably keep to what you have here and leave the other stuff to a future PR. This is already a nice improvement. Do you agree? (I will still perform a review.)