@richardotis 2018-06-03
As you suspected, pre-padding the points array would not be a safe optimization. The underlying raw C function has the size of the input inner dimension hard-coded in the pointer arithmetic for the dof array.