@bocklund 2018-06-05
My understanding is that we have to reshape based on the maximum tieline vertices no matter what, so we'll never hit that code. I decided to remove it.