@bocklund 2018-06-05
Merging this. For future reference line profiling has pointed out the following hotspots that could be optimized in the future (in order of time, time in seconds for order of magnitude gauge). Even after these changes, `_compute_phase_values` takes almost the entire time of `calculate` after the cache is warm.

### `calculate`
We can avoid all of the **Model**-labeled code by passing our own Models and callables. The ** cached**-labeled code is cached and only needs to be run once where the parameters are the same (I'm not so worried about fixing these immedately because they don't slow down `equilibrium`). 

1. (line 408) `phase_ds = _compute_phase_values(nonvacant_components, str_statevar_dict, points, phase_record, output, maximum_internal_dof, broadcast=broadcast, largest_energy=float(largest_energy), fake_points=fp)` 44s
2.  (line 409) `points = _sample_phase_constitution(phase_name, phase_obj.constituents, sublattice_dof, comps, tuple(variables), sampler_dict[phase_name] or point_sample, fixedgrid_dict[phase_name], pdens_dict[phase_name])` 35s - **cached**
3. (line 405) `phase_record = PhaseRecord_from_cython(comps, list(statevar_dict.keys()) + variables, np.array(dbf.phases[phase_name].sublattices, dtype=np.float), param_values,  comp_sets[phase_name], None, None, mass_dict[phase_name], None)` 19s - triggers JIT, **cached**
4. (line 385) `undefs = list(out.atoms(Symbol) - out.atoms(v.StateVariable))` 78s - **Model**
5. (line 387) `out = out.xreplace({undef: float(0)})` 0.5s - **Model**
6. (line 389) `comp_sets[phase_name] = build_functions(out, list(statevar_dict.keys()) + variables, include_obj=True, include_grad=False, parameters=param_symbols)` 0.2s - **Model**
7. (line 400) `mass_dict[phase_name] = [build_functions(mod.moles(el), list(statevar_dict.keys()) + variables, include_obj=True, include_grad=False, parameters=param_symbols)  for el in pure_elements]` 0.12s - **Model**



### `_compute_phase_values`
(this happens per phase in calculate) - it's mostly calculations and allocation.  

1. (line 188) `phase_record.obj(phase_output, dof)`  36s
2. (line 185) `dof = np.ascontiguousarray(np.concatenate((bc_statevars.T, pts), axis=1))` 3.7s
3. (line 190) `phase_record.mass_obj(phase_compositions[:,el_idx], dof, el_idx)` 1.8s
4. (line 206) `phase_names = np.full(points.shape[:-1], phase_record.phase_name, dtype='U'+str(len(phase_record.phase_name)))` 0.5s
5. (line 164) `statevars = np.meshgrid(*itertools.chain(statevar_dict.values(), [np.empty(points.shape[-2])]), sparse=True, indexing='ij')[:-1]` 0.3s
6. (line 183) `bc_statevars = np.ascontiguousarray([broadcast_to(x, points.shape[:-1]).reshape(-1) for x in statevars])` 0.2s